keystone-tools/requirements.txt
ed25519
pyelftools

